backend/requirements.txt
asgiref==3.11.1
Django==5.2.13
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
typing_extensions==4.15.0